+++ ClientSDK/C#/Source/ServerClient/ServerClientRequestCreateGame.cs
@@ -2,7 +2,7 @@
 
 public partial class ServerClient : NetClient
 {
-    public virtual void RequestCreateGame(string password, int mode, int mapid, int maxPlayers, float evaluation)
+    public virtual void RequestCreateGame(string password, int mode, int mapid, int maxPlayers, int evaluation)
     {
         mRequestCreateGame.password = password;
         mRequestCreateGame.mode = mode;
@@ -12,7 +12,7 @@ public partial class ServerClient : NetClient
         SendProto(ProtoGameClient.REQUEST_MSG.CREATE_GAME, mRequestCreateGame);
     }
 
-    public virtual void RequestCreateGame(string password, int gameid, int mode, int mapid, int maxPlayers, float evaluation)
+    public virtual void RequestCreateGame(string password, int gameid, int mode, int mapid, int maxPlayers, int evaluation)
     {
         mRequestCreateGame.password = password;
         mRequestCreateGame.gameid = gameid;
